fix: more functions work in api._
